Making create endpoint to return created customer

+++ JdbcExample/JdbcApplication/src/main/java/com/harishkannarao/jdbc/domain/DeleteCustomerRequestDto.java
@@ -1,4 +1,4 @@
 package com.harishkannarao.jdbc.domain;
 
-public record DeleteCustomerRequestDto(Long id) {
+public record DeleteCustomerRequestDto(Integer id) {
 }
